{{ cookiecutter.__project_slug }}/pyproject.toml
[project]
name = "{{ cookiecutter.__project_slug }}"
authors = [{name = "{{ cookiecutter.author_name }}", email = "{{ cookiecutter.author_email }}"}]
description = "{{ cookiecutter.project_description }}"
version = "0.0.0"
requires-python = ">=3.12"
dependencies = [
    "Django>=5.1",
    "django-allauth>=65.2",
    "huey>=2.5",
    "whitenoise>=6.8",
    "django-rich>=1.13",
    "ipython>=8.30",
    "django-environ>=0.11.2",
    "django-components>=0.114",
]

[tool.uv]
dev-dependencies = [
    "django-browser-reload",
    "django-debug-toolbar",
    "django-simple-factory",
    "ruff",
    "ipdb",
    "djade",
    "coverage",
]

[tool.ruff.isort]
force-single-line = true
